Utilility Consolidation + Cogs Added
